+++ python/src/opca/commands/crl/__init__.py
@@ -1,7 +1,0 @@
-# opca/commands/crl/__init__.py
-
-from .register import register
-
-__all__ = [
-    "register",
-]