+++ python/src/opca/commands/vault/__init__.py
@@ -1,5 +1,0 @@
-# opca/commands/vault/__init__.py
-
-from .register import register
-
-__all__ = ["register"]