+++ python/src/opca/models/__init__.py
@@ -1,7 +1,0 @@
-# opca/models/__init__.py
-
-from .app import App
-
-__all__ = [
-    "App",
-]